Initial commit
This commit is for an incomplete and unresponsive website. The top of the section needs CSS modification. The navBar is not properly sticking to the top and stays at the top for only one page scroll. I have to change the display for the contents of the webpage from flex to the grid display to properly align the adBanner and the mainContent of the webPage. The editorial id section has its columns not properly aligned so I have to fix it in future commits. I have to add sliders also in the future commits to the opinions section of the webpage. Only the entertainment section is more properly finished.